+++ packages/launcher/src/views/ModView/InstallView/InstallFromLocalForm.tsx
@@ -3,7 +3,6 @@ import { installModFromUrl } from '@/services/modStore/install';
 import Close from '@mui/icons-material/Close';
 import CloudUpload from '@mui/icons-material/CloudUploadRounded';
 import {
-    alpha,
     Button,
     Dialog,
     DialogActions,
@@ -81,7 +80,7 @@ export function InstallFromLocalForm() {
                 PaperProps={{
                     sx: {
                         backgroundImage: 'none',
-                        bgcolor: ({ palette }) => alpha(palette.popup.background, 0.8)
+                        bgcolor: ({ palette }) => palette.popup.background
                     },
                     component: 'form',
                     onSubmit: (event: React.FormEvent<HTMLFormElement>) => {